[FEAT] 게임 - log 저장

+++ BE/worldy/src/main/java/com/ssafy/worldy/model/game/dto/GameResultDto.java
@@ -1,5 +1,6 @@
 package com.ssafy.worldy.model.game.dto;
 
+import com.ssafy.worldy.model.game.entity.GameResult;
 import lombok.Getter;
 import lombok.Setter;
 
@@ -8,8 +9,9 @@ import java.util.List;
 @Getter
 @Setter
 public class GameResultDto {
-    List<Integer> first; // 1등이 보유한 나라들
-    List<Integer> second; // 1등이 보유한 나라들
-    List<Integer> third; // 1등이 보유한 나라들
-    List<Integer> fourth; // 1등이 보유한 나라들
+    List<Long> first; // 1등이 보유한 나라들
+    List<Long> second; // 1등이 보유한 나라들
+    List<Long> third; // 1등이 보유한 나라들
+    List<Long> fourth; // 1등이 보유한 나라들
+
 }
